tmux: Various updates
Main change was to set the "r" bind to source the correct tmux config
file at ~/.config/tmux/tmux.conf instead of ~/.tmux.conf as the install
script symlinks to the former.

This change also adjusts the status line a bit, just changing some
colors. As well, the full "set-option" commmand is used instead of the
"set" shorthand.