neovim: Fixed keybinding that broke copying to the system clipboard in tmux
85b4871 added the abliity to yank text from neovim's "+y register into
the system's clipboard using tmux, even over an SSH connection. This
keybinding broke that functinalilty, so in order to restore this useful
feature, the keybinding was removed.

85b4871 also added the feature that *any* yank from neovim running
within a tmux session copied the text to the system clipboard, but this
doesn't appear to be working any more. The only thing that's changed in
my environment is that I'm using a newer build of neovim from a commit
on the master branch made on 2023-01-01 (specifically
c590641febf4d03e89c46f8e7ef4c3fb2a455520). Perhaps something changed
within neovim that renders the autocommand added in 85b4871 now
obsolete, but it's not a big issue as long as yanking to the "+y
register continues to work within tmux.