neovim: Completely overhauled configuration
The configuration has been completely overhauled to be more standardized
and easier to read and update, especially with how LSPs are configured.
Some major changes include:

    - :PackerCompile is automatically executed whenever the neovim
      configuration is updated
    - Mason is used to manage language servers, which installs servers
      underneath it's own directory in ~/.local/share/nvim/mason rather
      than having to manually install language servers separately
      (though you still need package managers like NPM installed for
      some servers)
    - Treesitter grammars are automatically installed when a filetype is
      opened for the first time
    - The status line is now using lualine rather than airline
    - bufferline is now used for displaying tabs

  In addition, there are now a couple environment variables that can be
  set to configure the environment:

    - NVIM_FEATURE_LSP  :  Set to a falsy value to disable LSP globally
    - NVIM_LSP_SERVERS  :  Ensure that particular language servers are always installed
    - NVIM_COLORSCHEME  :  Specify the colorsheme to use without modifying any files

An example of using NVIM_LSP_SERVERS to ensure that language servers for
TypeScript, Bash, and JSON are always installed is as follows:

  export NVIM_LSP_SERVERS="tsserver,bashls,jsonls"

To specify the colorscheme, the value of NVIM_COLORSCHEME must be in the
following format:

  <plugin>/<theme_name>

This is to allow specifying a theme from a particular plugin in the
event that two or more plugins register a theme with the same name (for
example there *a lot* of onedark themes out there). An example of
NVIM_COLORSCHEME to use the minimal-base16 theme from
yazeed1s/minimal.nvim is as follows:

  export NVIM_COLORSCHEME="yazeed1s/minimal.nvim:minimal-base16"