feat(update-neovim.sh): Specify options via flags
In addition to specifying options via environment variables, options can
be specified by flags. Usage:

    -b, --branch=BRANCH        Specify the neovim branch to build
    -c, --num-commits=NUM      Specify the last NUM commits to show after fetching neovim updates
    -j, --jobs=NUM             Allow NUM jobs at once
    -n, --nice=NUM             Specify nice adjustment
    -p, --prefix=PATH          Directory prefix where neovim will be installed
    -s, --source=PATH          Directory path where neovim source is located

If the command line flag is specified, that value is used instead of the
environment variable (if present). If neither a command line flag *and*
environment variable is specified, then a default value is used. The
defaults are as follows:

    Branch:             master
    Build niceness:     +15
    Number of commits:  15
    Jobs:               <number of logical CPU cores from nproc/sysctl>
    Build prefix:       /usr/local/Cellar/neovim
    Source directory:   /usr/local/src/neovim