chore(neovim): Ignore more messages that would've been displayed via noice
jtdls will for some reason emit a "nil" and "ServiceReady" message when
the LSP server completes initializing. This change simply skips
displaying those messages so the annoying notification that noice
displays won't show. As a side effect, this could hide valid,
non-related messages that contain these strings, but it doesn't seem to
be much of an issue (for now).