fix(bin): Fixed semver compare in update-asdf-package
The previous algorithm didn't handle minor and patch versions correctly
so it incorrectly thought, for example, that "1.0.35" was greater than
"1.1.0".

The algorithm has been changed to a more crude version where the
versions are split into their major, minor, and patch versions via regex
and comparing each part to one another:

    Given two semvers ("latest" and "current"):

    - Return TRUE if latest major is greater then current major
    - Return TRUE If latest minor is greater than current minor
    - Return TRUE if latest minor is equal to current minor *and* latest
      patch is greater than current patch
    - Otherwise return FALSE because latest must either be less than or
      equal to current

The algorithm also assumes that if a version has a missing patch
version, then the patch version must be zero (i.e. "1.1" -> "1.1.0").