fix(install-asdf.sh): Support linux x86_64
Go uses "amd64" for x86_64, which matches the asset filename asdf uses
for its x86_64 artifacts.

Also, GNU mktemp has different flags than BSD mktemp (which macOS uses),
so when creating the temp directory for the asdf download, the flags for
GNU mktemp is tried first but if that command fails then the equivalent
BSD mktemp flags are used.